% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{bootwar}
\alias{bootwar}
\title{Bootwar Shiny App}
\usage{
bootwar()
}
\value{
A Shiny application object. Running this function will launch the Shiny app
in the user's default web browser.
}
\description{
Launches a Shiny application for the Bootwar card game.
The app allows users to play a card game where they can analyze the game
results using nonparametric bootstrap test with pooled resampling methods.
}
\details{
The Bootwar card game is a bootstrap variation of the card game War. The
Bootwar application has options to select different modes ('t' for
independent t-test and 'pt' for paired t-test) and decks.
Players can use a standard 52 card deck and they can also input a custom
anonymous function to generate a deck. The app will let users deal cards,
play the game, and then score and analyze results using nonparametric
bootstrap test with pooled resampling methods. The game is designed to help
users gain greater intuition on nonparametric bootstrap test with pooled
resampling methods; as such, players are encouraged to experiment with
different confidence levels, number of rounds, number of bootstrap resamples,
and custom decks.
}
\examples{
if(interactive()){
  bootwar()
}

}
