% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deck}
\alias{deck}
\title{Deck of Cards}
\format{
\subsection{\code{deck}}{

A data frame with 52 rows and 4 columns:
\describe{
\item{rank}{A factor representing card rank taking values 2 - A}
\item{suit}{A card suit with ranked order Club (C), Diamond (D), Heart (H), and Spade (S)}
\item{card}{A card}
\item{value}{A card value ranging from 2.00 (2C) to 14.75 (AS)}
}
}
}
\source{
\if{html}{\out{
<Standard Deck of Playing Cards>
}}
}
\usage{
deck
}
\description{
A 52 card deck of playing cards with suit ranking.
}
\keyword{datasets}
